// src/app/models/animal.interface.ts
export interface Animal{
    id?: string;
    nombre: string;
    raza: string;
    nacimiento: string;
    sexoAnimal: string;
    numInt: number;
    siiniga: string;
    padre: string;
    madre: string;
    numPierna: number;
    //Enfermedad
    //esta enlazado con medicina
    estado: string;
    //Lugar
    ubicacion: string;
    observaciones: string;
    //tiempo: string;
    fechadeTransferencia: string;
    vacia: string;
}